Backend configurado y listo

+++ Backend/menu/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 from .models import Ingredient, ProductIngredient, Meal, Drink, Dessert, Snack, Product
+from django.db import transaction
 
 # Serializer to register a new ingredient
 class IngredientSerializer(serializers.ModelSerializer):
@@ -96,6 +97,7 @@ class ProductWriteSerializer(serializers.ModelSerializer):
         
         return data
     
+    @transaction.atomic
     def create(self, validated_data):
         ingredients_data = validated_data.pop('ingredients', [])
         
@@ -123,6 +125,7 @@ class ProductWriteSerializer(serializers.ModelSerializer):
             
         return product
     
+    @transaction.atomic
     def update(self, instance, validated_data):
         ingredients_data = validated_data.pop('ingredients', None)
         
@@ -130,6 +133,9 @@ class ProductWriteSerializer(serializers.ModelSerializer):
         drink_data = validated_data.pop('drink', None)
         dessert_data = validated_data.pop('dessert', None)
         snack_data = validated_data.pop('snack', None)
+        
+        old_type = instance.product_type
+        new_type = validated_data.get('product_type', old_type)
         
         # Update basic fields
         for attr, value in validated_data.items():
@@ -139,26 +145,40 @@ class ProductWriteSerializer(serializers.ModelSerializer):
         # Ingredients
         if ingredients_data is not None:
             instance.product_ingredients.all().delete()
+            
+            ProductIngredient.objects.bulk_create([
+                ProductIngredient(
+                    product=instance,
+                    **item
+                )
+                for item in ingredients_data
+            ])
+
+            
+            '''
             for item in ingredients_data:
                 ProductIngredient.objects.create(product=instance, **item)
+            '''
         
-        Meal.objects.filter(product=instance).delete()
-        Drink.objects.filter(product=instance).delete()
-        Dessert.objects.filter(product=instance).delete()
-        Snack.objects.filter(product=instance).delete()
+        # Delete subtype only if type changes
+        if old_type != new_type:
+            Meal.objects.filter(product=instance).delete()
+            Drink.objects.filter(product=instance).delete()
+            Dessert.objects.filter(product=instance).delete()
+            Snack.objects.filter(product=instance).delete()
         
-        # Specific type
+        # Recreate subtype if provided
         if instance.product_type == 'meal' and meal_data:
-            Meal.objects.update_or_create(product=instance, defaults=meal_data)
+            Meal.objects.create(product=instance, **meal_data)
         
         elif instance.product_type == 'drink' and drink_data:
-            Drink.objects.update_or_create(product=instance, defaults=drink_data)
+            Drink.objects.create(product=instance, **drink_data)
             
         elif instance.product_type == 'dessert' and dessert_data:
-            Dessert.objects.update_or_create(product=instance, defaults=dessert_data)
+            Dessert.objects.create(product=instance, **dessert_data)
         
         elif instance.product_type == 'snack' and snack_data:
-            Snack.objects.update_or_create(product=instance, defaults=snack_data)
+            Snack.objects.create(product=instance, **snack_data)
         
         return instance
